requirements.txt
streamlit
openai
pandas
openpyxl
Office365-REST-Python-Client
xlsxwriter
python-dotenv
